+++ charts/datadog-operator/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## 2.24.0-dev.4
+
+* Update Datadog Operator chart for 1.28.0-rc.3.
+
+## 2.24.0-dev.3
+
+* Update Datadog Operator chart for 1.28.0-rc.2.
+
 ## 2.24.0-dev.2
 
 * Same as 2.24.0-dev.1 including 2.23.2 changelog entry to resume development.